+++ Naheulbook.Data/Models/OriginRestrict.cs
@@ -2,11 +2,11 @@ namespace Naheulbook.Data.Models
 {
     public class OriginRestrict
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Text { get; set; }
         public string Flags { get; set; }
 
-        public long OriginId { get; set; }
+        public int OriginId { get; set; }
         public Origin Origin { get; set; }
     }
 }
